+++ src/components/Card/Card.jsx
@@ -4,15 +4,20 @@ import React, { Component } from 'react';
 export class Card extends Component{
     render(){
         return (
-            <div className={"card "+this.props.cardClass}>
-                <div className="header">
+            <div className={"card"+(this.props.plain ? " card-plain":"")}>
+                <div className={"header"
+                    + (this.props.hCenter ? " text-center":"")}>
                     <h4 className="title">{this.props.title}</h4>
                     <p className="category">{this.props.category}</p>
                 </div>
-                <div className={"content "+this.props.contentClass}>
-                    <div id={this.props.id} className={this.props.classes}>
-                        {this.props.content}
-                    </div>
+                <div className={"content"
+                    + (this.props.ctAllIcons ? " all-icons":"")
+                    + (this.props.ctTableFullWidth ? " table-full-width":"")
+                    + (this.props.ctTableResponsive ? " table-responsive":"")
+                    + (this.props.ctTableUpgrade ? " table-upgrade":"")}>
+
+                    {this.props.content}
+
                     <div className="footer">
                         {this.props.legend}
                         {this.props.stats != null ? <hr />:""}