+++ src/components/Footer/Footer.jsx
@@ -1,11 +1,11 @@
 import React, {Component} from 'react';
-
+import { Grid } from 'react-bootstrap';
 
 class Footer extends Component {
 	render() {
 		return (
             <footer className="footer">
-                <div className="container-fluid">
+                <Grid>
                     <nav className="pull-left">
                         <ul>
                             <li>
@@ -33,7 +33,7 @@ class Footer extends Component {
                     <p className="copyright pull-right">
                         &copy; {(new Date()).getFullYear()} <a href="http://www.creative-tim.com">Creative Tim</a>, made with love for a better web
                     </p>
-                </div>
+                </Grid>
             </footer>
 		);
 	}