+++ src/components/Header/Header.jsx
@@ -1,7 +1,9 @@
-import React, {Component} from 'react';
+import React, { Component } from 'react';
 import { Navbar } from 'react-bootstrap';
 
 import HeaderLinks from './HeaderLinks.jsx';
+
+import appRoutes from 'routes/app.jsx';
 
 class Header extends Component{
     constructor(props){
@@ -28,16 +30,41 @@ class Header extends Component{
         };
         document.body.appendChild(node);
     }
+    getBrand(){
+        var name;
+        appRoutes.map((prop,key) => {
+            if(prop.collapse){
+                 prop.views.map((prop,key) => {
+                    if(prop.path === this.props.location.pathname){
+                        name = prop.name;
+                    }
+                    return null;
+                })
+            } else {
+                if(prop.redirect){
+                    if(prop.path === this.props.location.pathname){
+                        name = prop.name;
+                    }
+                }else{
+                    if(prop.path === this.props.location.pathname){
+                        name = prop.name;
+                    }
+                }
+            }
+            return null;
+        })
+        return name;
+    }
     render(){
         return (
             <Navbar fluid>
                 <Navbar.Header>
                     <Navbar.Brand>
-                        <a href="#pablo">{this.props.location.pathname.charAt(1).toUpperCase() + this.props.location.pathname.slice(2)}</a>
+                        <a href="#pablo">{this.getBrand()}</a>
                     </Navbar.Brand>
                     <Navbar.Toggle onClick={this.mobileSidebarToggle}/>
                 </Navbar.Header>
-                <Navbar.Collapse id="header_navbar">
+                <Navbar.Collapse>
                     <HeaderLinks />
                 </Navbar.Collapse>
             </Navbar>