+++ src/components/StatsCard/StatsCard.jsx
@@ -1,24 +1,24 @@
 import React, { Component } from 'react';
-
+import { Row, Col } from 'react-bootstrap';
 
 export class StatsCard extends Component{
     render(){
         return (
             <div className="card card-stats">
                 <div className="content">
-                    <div className="row">
-                        <div className="col-xs-5">
+                    <Row>
+                        <Col xs={5}>
                             <div className="icon-big text-center icon-warning">
                                 {this.props.bigIcon}
                             </div>
-                        </div>
-                        <div className="col-xs-7">
+                        </Col>
+                        <Col xs={7}>
                             <div className="numbers">
                                 <p>{this.props.statsText}</p>
                                 {this.props.statsValue}
                             </div>
-                        </div>
-                    </div>
+                        </Col>
+                    </Row>
                     <div className="footer">
                         <hr />
                         <div className="stats">