+++ src/elements/CustomCheckbox/CustomCheckbox.jsx
@@ -4,7 +4,7 @@ class CustomCheckbox extends Component {
     constructor(props){
         super(props);
             this.state = {
-                is_checked: props.isChecked
+                is_checked: (props.isChecked ? true:false)
             };
         this.handleClick = this.handleClick.bind(this);
     }
@@ -12,10 +12,10 @@ class CustomCheckbox extends Component {
         this.setState({is_checked: !this.state.is_checked});
     }
     render() {
-        const { isChecked, number, label,...rest } = this.props;
-
+        const { isChecked, number, label, inline,...rest } = this.props;
+        const classes = inline !== undefined ? "checkbox checkbox-inline":"checkbox";
         return (
-            <div className="checkbox">
+            <div className={classes}>
                 <input id={number} type="checkbox" onChange={this.handleClick} checked={this.state.is_checked} {...rest}/>
                 <label htmlFor={number}>{label}
                 </label>