+++ src/views/Dashboard/Dashboard.jsx
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
 import ChartistGraph from 'react-chartist';
+import { Grid, Row, Col } from 'react-bootstrap';
 
 
 import {Card} from 'components/Card/Card.jsx';
@@ -36,9 +37,9 @@ class Dashboard extends Component {
     render() {
         return (
             <div className="content">
-                <div className="container-fluid">
-                    <div className="row">
-                        <div className="col-lg-3 col-sm-6">
+                <Grid fluid>
+                    <Row>
+                        <Col lg={3} sm={6}>
                             <StatsCard
                                 bigIcon={<i className="pe-7s-server text-warning"></i>}
                                 statsText="Capacity"
@@ -46,8 +47,8 @@ class Dashboard extends Component {
                                 statsIcon={<i className="fa fa-refresh"></i>}
                                 statsIconText="Updated now"
                             />
-                        </div>
-                        <div className="col-lg-3 col-sm-6">
+                        </Col>
+                        <Col lg={3} sm={6}>
                             <StatsCard
                                 bigIcon={<i className="pe-7s-wallet text-success"></i>}
                                 statsText="Revenue"
@@ -55,8 +56,8 @@ class Dashboard extends Component {
                                 statsIcon={<i className="fa fa-calendar-o"></i>}
                                 statsIconText="Last day"
                             />
-                        </div>
-                        <div className="col-lg-3 col-sm-6">
+                        </Col>
+                        <Col lg={3} sm={6}>
                             <StatsCard
                                 bigIcon={<i className="pe-7s-graph1 text-danger"></i>}
                                 statsText="Errors"
@@ -64,8 +65,8 @@ class Dashboard extends Component {
                                 statsIcon={<i className="fa fa-clock-o"></i>}
                                 statsIconText="In the last hour"
                             />
-                        </div>
-                        <div className="col-lg-3 col-sm-6">
+                        </Col>
+                        <Col lg={3} sm={6}>
                             <StatsCard
                                 bigIcon={<i className="fa fa-twitter text-info"></i>}
                                 statsText="Followers"
@@ -73,23 +74,25 @@ class Dashboard extends Component {
                                 statsIcon={<i className="fa fa-refresh"></i>}
                                 statsIconText="Updated now"
                             />
-                        </div>
-                    </div>
-                    <div className="row">
-                        <div className="col-md-8">
+                        </Col>
+                    </Row>
+                    <Row>
+                        <Col md={8}>
                             <Card
                                 statsIcon="fa fa-history"
                                 id="chartHours"
-                                classes="ct-chart"
                                 title="Users Behavior"
                                 category="24 Hours performance"
                                 stats="Updated 3 minutes ago"
                                 content={
-                                    <ChartistGraph
-                                        data={dataSales}
-                                        type="Line"
-                                        options={optionsSales}
-                                        responsiveOptions={responsiveSales}/>
+                                    <div className="ct-chart">
+                                        <ChartistGraph
+                                            data={dataSales}
+                                            type="Line"
+                                            options={optionsSales}
+                                            responsiveOptions={responsiveSales}
+                                        />
+                                    </div>
                                     }
                                 legend={
                                     <div className="legend">
@@ -97,17 +100,17 @@ class Dashboard extends Component {
                                     </div>
                                 }
                             />
-                        </div>
-                        <div className="col-md-4">
+                        </Col>
+                        <Col md={4}>
                             <Card
                                 statsIcon="fa fa-clock-o"
-                                id="chartPreferences"
-                                classes="ct-chart ct-perfect-fourth"
                                 title="Email Statistics"
                                 category="Last Campaign Performance"
                                 stats="Campaign sent 2 days ago"
                                 content={
-                                    <ChartistGraph data={dataPie} type="Pie"/>
+                                    <div id="chartPreferences" className="ct-chart ct-perfect-fourth">
+                                        <ChartistGraph data={dataPie} type="Pie"/>
+                                    </div>
                                 }
                                 legend={
                                     <div className="legend">
@@ -115,25 +118,26 @@ class Dashboard extends Component {
                                     </div>
                                 }
                             />
-                        </div>
-                    </div>
+                        </Col>
+                    </Row>
 
-                    <div className="row">
-                        <div className="col-md-6">
+                    <Row>
+                        <Col md={6}>
                             <Card
                                 id="chartActivity"
-                                classes="ct-chart"
                                 title="2014 Sales"
                                 category="All products including Taxes"
                                 stats="Data information certified"
                                 statsIcon="fa fa-check"
                                 content={
-                                    <ChartistGraph
-                                        data={dataBar}
-                                        type="Bar"
-                                        options={optionsBar}
-                                        responsiveOptions={responsiveBar}
-                                    />
+                                    <div className="ct-chart">
+                                        <ChartistGraph
+                                            data={dataBar}
+                                            type="Bar"
+                                            options={optionsBar}
+                                            responsiveOptions={responsiveBar}
+                                        />
+                                    </div>
                                 }
                                 legend={
                                     <div className="legend">
@@ -141,25 +145,26 @@ class Dashboard extends Component {
                                     </div>
                                 }
                             />
-                        </div>
+                        </Col>
 
-                        <div className="col-md-6">
+                        <Col md={6}>
                             <Card
-                                classes="table-full-width"
                                 title="Tasks"
                                 category="Backend development"
                                 stats="Updated 3 minutes ago"
                                 statsIcon="fa fa-history"
                                 content={
-                                    <table className="table">
-                                        <Tasks />
-                                    </table>
+                                    <div className="table-full-width">
+                                        <table className="table">
+                                            <Tasks />
+                                        </table>
+                                    </div>
                                 }
                             />
-                        </div>
-                    </div>
+                        </Col>
+                    </Row>
 
-                </div>
+                </Grid>
             </div>
         );
     }