+++ src/views/Typography/Typography.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-
+import { Grid, Row, Col } from 'react-bootstrap';
 
 import Card from 'components/Card/Card.jsx'
 
@@ -7,9 +7,9 @@ class Typography extends Component {
     render() {
         return (
             <div className="content">
-                <div className="container-fluid">
-                    <div className="row">
-                        <div className="col-md-12">
+                <Grid fluid>
+                    <Row>
+                        <Col md={12}>
                             <Card
                                 title="Light Bootstrap Table Heading"
                                 category="Created using Roboto Font Family"
@@ -83,9 +83,9 @@ class Typography extends Component {
                                     </div>
                                 }
                             />
-                        </div>
-                    </div>
-                </div>
+                        </Col>
+                    </Row>
+                </Grid>
             </div>
         );
     }