+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## [1.1.0] 2017-10-30
+### Bug fixing
+- Changed property of `isChecked` in `CustomCheckbox`
+- Removed hover state for checkboxes and radios
+- Moved `NotificationSystem` outside of `.main-panel` (for responsive issues)
+- All `div`'s with Bootstrap classes (`.container`,`.container-fluid`,`.row`,`.col-*`) were changed in React-Bootstrap components
+- Changed `Notifications Places` buttons from buttons with `className` to our own button component `CustomButton`
+- Added property `round` in `CustomButton` (for the button class `.btn-round`)
+- Added dynamic routes
+
+
 ## [1.0.0] 2017-09-20
 ### Original Release
 - Added React-Bootstrap as base framework